feat: rename sampling_mode config
