+++ examples/sft/config/libero_sft_dreamzero_5b.yaml
@@ -37,7 +37,7 @@ data:
   prefetch_factor: 8
   video_tolerance_s: 0.1
   parquet_cache_size: 512
-  sampling_mode: sharded # dense
+  sampling_mode: multi_anchor # fixed_window
 
 algorithm:
   adv_type: gae
@@ -48,6 +48,7 @@ actor:
   micro_batch_size: 32
   global_batch_size: 256
   seed: 42
+
   model:
     model_type: "dreamzero"
     precision: fp32
@@ -60,15 +61,8 @@ actor:
     # Dataset-generated metadata.json for normalisation statistics.
     # Must be explicitly specified unless it exists at model_path/experiment_cfg/metadata.json.
     metadata_json_path: /mnt/project_rlinf/guozhen/test_wam/test_dataset/metadata.json
-
-    max_seq_len: 512
-    max_action_dim: 32
-    max_state_dim: 64
-    state_horizon: 1
     action_horizon: 16
     num_chunks: 4
-    num_video_frames: 33
-
     num_action_per_block: 16
     target_video_height: 160
     target_video_width: 320
