+++ rlinf/data/datasets/dreamzero/data_transforms/__init__.py
@@ -12,10 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""DreamZero ``ComposedModalityTransform`` built from per-embodiment modules."""
-
-from __future__ import annotations
-
+import ast
 import json
 from pathlib import Path
 from typing import Any
@@ -23,27 +20,80 @@ from typing import Any
 from groot.vla.data.schema import DatasetMetadata
 from groot.vla.data.transform.base import ComposedModalityTransform
 
-from rlinf.models.embodiment.dreamzero.data_transforms.libero_sim import (
+from rlinf.data.datasets.dreamzero.data_transforms.base import DreamZeroEmbodimentTransform
+from rlinf.data.datasets.dreamzero.data_transforms.libero_sim import (
     LiberoSimDataTransform,
 )
-from rlinf.models.embodiment.dreamzero.data_transforms.oxe_droid import (
+from rlinf.data.datasets.dreamzero.data_transforms.oxe_droid import (
     OxeDroidDataTransform,
 )
 
-# Projector indices for built-in embodiments (must match DreamTransform / collate).
+_EMBODIMENT_REGISTRY: dict[str, type[DreamZeroEmbodimentTransform]] = {
+    LiberoSimDataTransform.TAG: LiberoSimDataTransform,
+    OxeDroidDataTransform.TAG: OxeDroidDataTransform,
+}
+
 DEFAULT_EMBODIMENT_TAG_MAPPING: dict[str, dict[str, int]] = {
-    "libero_sim": {"libero_sim": 21},
-    "oxe_droid": {"oxe_droid": 17},
+    tag: dict(cls.DEFAULT_TAG_MAPPING)
+    for tag, cls in _EMBODIMENT_REGISTRY.items()
 }
 
 __all__ = [
-    "DEFAULT_EMBODIMENT_TAG_MAPPING",
     "build_dreamzero_composed_transform",
     "collect_dreamzero_dataset_keys",
     "embodiment_tag_mapping_for_embodiment",
-    "language_keys_for_embodiment",
     "load_dreamzero_dataset_metadata",
+    "format_training_prompt",
+    "normalize_instruction_text",
 ]
+
+
+def _require_embodiment(tag: str) -> type[DreamZeroEmbodimentTransform]:
+    try:
+        return _EMBODIMENT_REGISTRY[tag]
+    except KeyError:
+        raise ValueError(
+            f"Unsupported embodiment_tag {tag!r}. "
+            f"Built-in tags: {sorted(_EMBODIMENT_REGISTRY)}. "
+            "Register the class in _EMBODIMENT_REGISTRY."
+        ) from None
+
+
+def _language_keys_for_tag(tag: str) -> list[str]:
+    modality = _require_embodiment(tag).get_modality_config()
+    language_cfg = modality.get("language")
+    if language_cfg is None:
+        raise KeyError(f"Missing language ModalityConfig for {tag!r}")
+    return [str(k) for k in language_cfg.modality_keys]
+
+
+def normalize_instruction_text(raw: Any) -> str:
+    """Decode a dataset ``text`` field into a lowercase instruction string."""
+    if not isinstance(raw, str):
+        return str(raw).lower()
+    try:
+        parsed = ast.literal_eval(raw)
+        if isinstance(parsed, (list, tuple)):
+            return str(parsed[0]).lower()
+        return str(parsed).lower()
+    except (ValueError, SyntaxError, TypeError):
+        return raw.lower()
+
+
+def format_training_prompt(
+    instruction: str,
+    embodiment_id: int,
+    embodiment_tag_mapping: dict[str, int],
+) -> str:
+    """Wrap a task instruction with embodiment-specific multi-view layout text for T5."""
+    id_to_tag = {v: k for k, v in embodiment_tag_mapping.items()}
+    tag = id_to_tag.get(embodiment_id)
+    if tag is None:
+        raise ValueError(
+            f"Embodiment ID {embodiment_id} not found in embodiment_tag_mapping "
+            f"{embodiment_tag_mapping!r}."
+        )
+    return _require_embodiment(tag).format_training_prompt(instruction)
 
 
 def embodiment_tag_mapping_for_embodiment(
@@ -53,13 +103,7 @@ def embodiment_tag_mapping_for_embodiment(
     """Return embodiment tag -> projector id mapping for collate / DreamTransform."""
     if override is not None:
         return dict(override)
-    try:
-        return dict(DEFAULT_EMBODIMENT_TAG_MAPPING[tag])
-    except KeyError:
-        raise ValueError(
-            f"Unsupported embodiment_tag {tag!r}; set actor.model.embodiment_tag_mapping. "
-            f"Built-in tags: {list(DEFAULT_EMBODIMENT_TAG_MAPPING)}."
-        ) from None
+    return dict(_require_embodiment(tag).DEFAULT_TAG_MAPPING)
 
 
 def collect_dreamzero_dataset_keys(
@@ -74,25 +118,8 @@ def collect_dreamzero_dataset_keys(
         video_keys.extend(getattr(transform, "video_concat_order", []) or [])
         state_keys.extend(getattr(transform, "state_concat_order", []) or [])
         action_keys.extend(getattr(transform, "action_concat_order", []) or [])
-    language_keys = language_keys_for_embodiment(embodiment_tag)
+    language_keys = _language_keys_for_tag(embodiment_tag)
     return video_keys, state_keys, action_keys, language_keys
-
-
-def language_keys_for_embodiment(tag: str) -> list[str]:
-    """Return language modality keys from the embodiment transform config."""
-    if tag == "oxe_droid":
-        modality = OxeDroidDataTransform.get_modality_config()
-    elif tag == "libero_sim":
-        modality = LiberoSimDataTransform.get_modality_config()
-    else:
-        raise ValueError(f"Unsupported embodiment_tag {tag!r} for built-in DreamZero transforms. "
-        "Currently only oxe_droid and libero_sim are supported. "
-        "To add a new embodiment, create a module under "
-        "rlinf.models.embodiment.dreamzero.data_transforms.")
-    language_cfg = modality.get("language")
-    if language_cfg is None:
-        raise KeyError(f"Missing language ModalityConfig for {tag!r}")
-    return [str(k) for k in language_cfg.modality_keys]
 
 
 def load_dreamzero_dataset_metadata(cfg: Any) -> DatasetMetadata:
@@ -133,49 +160,12 @@ def build_dreamzero_composed_transform(
 ) -> ComposedModalityTransform:
     """Construct ``ComposedModalityTransform`` for the current ``embodiment_tag``."""
     tag = cfg.embodiment_tag
+    cls = _require_embodiment(tag)
     embodiment_tag_mapping = embodiment_tag_mapping_for_embodiment(
         tag, cfg.get("embodiment_tag_mapping")
     )
-
-    if tag == "oxe_droid":
-        return OxeDroidDataTransform.get_transform(
-            tokenizer_path=tokenizer_path,
-            state_horizon=cfg.get("state_horizon", 1),
-            action_horizon=cfg.get("action_horizon", 24),
-            max_state_dim=cfg.get("max_state_dim", 64),
-            max_action_dim=cfg.get("max_action_dim", 32),
-            max_length=cfg.get("max_seq_len", 512),
-            default_instruction=cfg.get(
-                "default_instruction", "Perform the default behavior."
-            ),
-            language_dropout_prob=cfg.get("language_dropout_prob", 0.0),
-            always_use_default_instruction=cfg.get(
-                "always_use_default_instruction", False
-            ),
-            embodiment_tag_mapping=embodiment_tag_mapping,
-        )
-
-    if tag == "libero_sim":
-        return LiberoSimDataTransform.get_transform(
-            tokenizer_path=tokenizer_path,
-            state_horizon=cfg.get("state_horizon", 1),
-            action_horizon=cfg.get("action_horizon", 16),
-            max_state_dim=cfg.get("max_state_dim", 64),
-            max_action_dim=cfg.get("max_action_dim", 32),
-            max_length=cfg.get("max_seq_len", 512),
-            default_instruction=cfg.get(
-                "default_instruction", "Perform the default behavior."
-            ),
-            language_dropout_prob=cfg.get("language_dropout_prob", 0.0),
-            always_use_default_instruction=cfg.get(
-                "always_use_default_instruction", False
-            ),
-            embodiment_tag_mapping=embodiment_tag_mapping,
-        )
-
-    raise ValueError(
-        f"Unsupported embodiment_tag {tag!r} for built-in DreamZero transforms. "
-        "Currently only oxe_droid and libero_sim are supported. "
-        "To add a new embodiment, create a module under "
-        "rlinf.models.embodiment.dreamzero.data_transforms."
+    return cls.get_transform(
+        tokenizer_path=tokenizer_path,
+        cfg=cfg,
+        embodiment_tag_mapping=embodiment_tag_mapping,
     )
