+++ rlinf/data/datasets/dreamzero/data_transforms/libero_sim.py
@@ -14,6 +14,7 @@
 
 from typing import Any
 
+import numpy as np
 from groot.vla.data.dataset.lerobot import ModalityConfig
 from groot.vla.data.transform.base import ComposedModalityTransform
 from groot.vla.data.transform.concat import ConcatTransform
@@ -28,31 +29,53 @@ from groot.vla.data.transform.video import (
     VideoToNumpy,
     VideoToTensor,
 )
-from groot.vla.model.dreamzero.transform.dreamzero_cotrain import DreamTransform
+from rlinf.data.datasets.dreamzero.data_transforms.dream_transform import DreamTransform
 
 
 _VIDEO_KEYS = [
-    "video.exterior_image_1_left",
-    "video.exterior_image_2_left",
-    "video.wrist_image_left",
+    "video.image",
+    "video.wrist_image",
 ]
-_STATE_KEYS = ["state.joint_position", "state.gripper_position"]
-_ACTION_KEYS = ["action.joint_position", "action.gripper_position"]
+_STATE_KEYS = ["state.state"]
+_ACTION_KEYS = ["action.actions"]
 
 _VIDEO_BACKEND = "torchvision"
 
+_TRAINING_PROMPT_PREFIX = "A multi-view video shows that a robot "
+_MULTIVIEW_LAYOUT = (
+    " The video is split into two horizontal views: the left view shows the "
+    "exterior camera and the right view shows the wrist camera. The robot "
+)
 
-class OxeDroidDataTransform:
-    """Provides modality config and composed transform for oxe_droid.
-    """
 
-    # ------------------------------------------------------------------
-    # Modality config
-    # ------------------------------------------------------------------
+class LiberoSimDataTransform:
+    """Provides modality config and composed transform for libero_sim."""
+
+    TAG = "libero_sim"
+    DEFAULT_TAG_MAPPING = {"libero_sim": 21}
+    DEFAULT_ACTION_HORIZON = 16
+
+    @staticmethod
+    def format_training_prompt(instruction: str) -> str:
+        """Build multi-view layout prompt for LIBERO (matches Groot collate template)."""
+        return _TRAINING_PROMPT_PREFIX + instruction + _MULTIVIEW_LAYOUT + instruction
+
+    @staticmethod
+    def concat_multiview_video(images: np.ndarray) -> np.ndarray:
+        """Horizontal concat: exterior (left) | wrist (right)."""
+        v, t, c, h, w = images.shape
+        if v < 2:
+            raise ValueError(
+                f"libero_sim expects at least 2 video views, got v={v} with shape {images.shape}"
+            )
+        concat_images = np.zeros((1, t, c, h, 2 * w), dtype=images.dtype)
+        concat_images[0, :, :, :, :w] = images[0]
+        concat_images[0, :, :, :, w:] = images[1]
+        return concat_images
 
     @staticmethod
     def get_modality_config() -> dict[str, ModalityConfig]:
-        """Return modality config dict for oxe_droid (25 video delta, 24 action delta)."""
+        """Return modality config dict for libero_sim (25 video delta, 24 action delta)."""
         return {
             "video": ModalityConfig(
                 delta_indices=list(range(25)),
@@ -69,11 +92,7 @@ class OxeDroidDataTransform:
             ),
             "language": ModalityConfig(
                 delta_indices=[0],
-                modality_keys=[
-                    "annotation.language.language_instruction",
-                    "annotation.language.language_instruction_2",
-                    "annotation.language.language_instruction_3",
-                ],
+                modality_keys=["annotation.task"],
             ),
             "lapa_action": ModalityConfig(
                 delta_indices=[0],
@@ -81,38 +100,46 @@ class OxeDroidDataTransform:
             ),
         }
 
-    # ------------------------------------------------------------------
-    # Composed transform
-    # ------------------------------------------------------------------
-
     @staticmethod
     def get_transform(
+        *,
         tokenizer_path: str,
-        state_horizon: int = 1,
-        action_horizon: int = 24,
-        max_state_dim: int = 64,
-        max_action_dim: int = 32,
-        max_length: int = 512,
-        default_instruction: str = "Perform the default behavior.",
-        language_dropout_prob: float = 0.0,
-        always_use_default_instruction: bool = False,
-        embodiment_tag_mapping: dict[str, int] | None = None,
+        cfg: Any,
+        embodiment_tag_mapping: dict[str, int],
     ) -> ComposedModalityTransform:
-        """Build the full ``ComposedModalityTransform`` chain for oxe_droid.
+        """Build the full ``ComposedModalityTransform`` chain for libero_sim."""
+        return LiberoSimDataTransform._build_composed_transform(
+            tokenizer_path=tokenizer_path,
+            state_horizon=int(cfg.get("state_horizon", 1)),
+            action_horizon=int(
+                cfg.get("action_horizon", LiberoSimDataTransform.DEFAULT_ACTION_HORIZON)
+            ),
+            max_state_dim=int(cfg.get("max_state_dim", 64)),
+            max_action_dim=int(cfg.get("max_action_dim", 32)),
+            max_length=int(cfg.get("max_seq_len", 512)),
+            default_instruction=str(
+                cfg.get("default_instruction", "Perform the default behavior.")
+            ),
+            language_dropout_prob=float(cfg.get("language_dropout_prob", 0.0)),
+            always_use_default_instruction=bool(
+                cfg.get("always_use_default_instruction", False)
+            ),
+            embodiment_tag_mapping=dict(embodiment_tag_mapping),
+        )
 
-        The chain is: VideoToTensor -> VideoCrop(0.95) -> VideoResize(176,320)
-        -> VideoColorJitter(0.3,0.4,0.5,0.08) -> VideoToNumpy ->
-        StateActionToTensor(state) -> StateActionTransform(state, q99) ->
-        StateActionToTensor(action) -> StateActionTransform(action, q99) ->
-        ConcatTransform -> DreamTransform.
-        """
-        if embodiment_tag_mapping is None:
-            from rlinf.models.embodiment.dreamzero.data_transforms import (
-                DEFAULT_EMBODIMENT_TAG_MAPPING,
-            )
-
-            embodiment_tag_mapping = DEFAULT_EMBODIMENT_TAG_MAPPING["oxe_droid"]
-
+    @staticmethod
+    def _build_composed_transform(
+        tokenizer_path: str,
+        state_horizon: int,
+        action_horizon: int,
+        max_state_dim: int,
+        max_action_dim: int,
+        max_length: int,
+        default_instruction: str,
+        language_dropout_prob: float,
+        always_use_default_instruction: bool,
+        embodiment_tag_mapping: dict[str, int],
+    ) -> ComposedModalityTransform:
         vk = list(_VIDEO_KEYS)
         state_k = list(_STATE_KEYS)
         action_k = list(_ACTION_KEYS)
@@ -123,8 +150,8 @@ class OxeDroidDataTransform:
             VideoResize(
                 apply_to=vk,
                 backend=_VIDEO_BACKEND,
-                height=176,
-                width=320,
+                height=256,
+                width=256,
                 interpolation="linear",
             ),
             VideoColorJitter(
@@ -139,18 +166,12 @@ class OxeDroidDataTransform:
             StateActionToTensor(apply_to=state_k),
             StateActionTransform(
                 apply_to=state_k,
-                normalization_modes={
-                    "state.joint_position": "q99",
-                    "state.gripper_position": "q99",
-                },
+                normalization_modes={"state.state": "q99"},
             ),
             StateActionToTensor(apply_to=action_k),
             StateActionTransform(
                 apply_to=action_k,
-                normalization_modes={
-                    "action.joint_position": "q99",
-                    "action.gripper_position": "q99",
-                },
+                normalization_modes={"action.actions": "q99"},
             ),
             ConcatTransform(
                 apply_to=[],
